Move unused code to Unused internal namespace
