+++ src/Imageflow/Fluent/ImageJob.cs
@@ -300,40 +300,6 @@ public class ImageJob : IDisposable
         }
     }
 
-#pragma warning disable CS0618 // Type or member is obsolete
-    private class StreamJsonSpanProvider : IJsonResponseProvider, IJsonResponseSpanProvider, IJsonResponse
-#pragma warning restore CS0618 // Type or member is obsolete
-    {
-        private readonly MemoryStream _ms;
-
-        public StreamJsonSpanProvider(int statusCode, MemoryStream ms)
-        {
-            ImageflowErrorCode = statusCode;
-            _ms = ms;
-        }
-        public void Dispose() => _ms.Dispose();
-        public Stream GetStream() => _ms;
-        public ReadOnlySpan<byte> BorrowBytes()
-        {
-            return _ms.TryGetBufferSliceAllWrittenData(out var slice) ? slice : _ms.ToArray();
-        }
-
-        public int ImageflowErrorCode { get; }
-        public string CopyString()
-        {
-            return BorrowBytes().Utf8ToString();
-        }
-
-        public JsonNode? Parse()
-        {
-            return BorrowBytes().ParseJsonNode();
-        }
-
-        public byte[] CopyBytes()
-        {
-            return BorrowBytes().ToArray();
-        }
-    }
 
     // private object BuildJsonWithPlaceholders()
     // {
@@ -803,3 +769,4 @@ public class ImageJob : IDisposable
         return MagicBytes.GetImageContentType(first12Bytes);
     }
 }
+
