+++ src/Imageflow/Internal.Helpers/StreamMemoryExtensionsPolyfills.cs
@@ -1,5 +1,7 @@
+#if !NETSTANDARD2_1_OR_GREATER
 using System.Buffers;
 using System.Runtime.InteropServices;
+#endif
 
 namespace Imageflow.Internal.Helpers;
