+++ src/Imageflow/Internal.Helpers/Unused/OutputDestinationWrapper.cs
